pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alma-bulk-tools"
version = "0.1.0"
description = "CLI-first ALMA archive bulk discovery, download, unpack, summarize, and indexing tools"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "alma-bulk-tools"}]
license = {text = "MIT"}
dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
alma-bulk = "alma_bulk_tools.cli:main"
alma-discover = "alma_bulk_tools.cli:main_discover"
alma-download = "alma_bulk_tools.cli:main_download"
alma-unpack = "alma_bulk_tools.cli:main_unpack"
alma-summarize = "alma_bulk_tools.cli:main_summarize"
alma-index-scan = "alma_bulk_tools.cli:main_scan"
alma-merge-index = "alma_bulk_tools.cli:main_merge_index"
alma-status = "alma_bulk_tools.cli:main_status"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["alma_bulk_tools*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
